requirements.txt
Django>=5.2.4
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5
django-extensions>=3.2.3
drf-spectacular>=0.27.0
python-dotenv>=1.0.0
Pillow>=10.1.0
psycopg2-binary>=2.9.7
django-debug-toolbar>=4.2.0
djangorestframework-simplejwt>=5.3.0 